Improved calorie tracker functionality and UI
Completed local storage for workouts, including adding, removing, and clearing.

Displays a message when there are no meals or workouts, and disables/updates the filter input accordingly.

Added filtering functionality that displays a message when no results are found.

Prompts the user for confirmation before resetting the tracker to prevent accidental data loss.